Eval: Store array index in EvaluationError::ArrayIndexOutOfBounds as i64
If a program accidentaly indexes using a negative value then we want
to report that value, therefore we can't use a 'usize' like before

+++ src/interpreter/ast.rs
@@ -64,7 +64,7 @@ pub enum ImmediateValue {
 #[derive(Debug, PartialEq, Clone)]
 pub enum LValue {
     Identifier(String),
-    IndexInArray(InternalArrayRepresentation, usize),
+    IndexInArray(InternalArrayRepresentation, i64),
 }
 
 #[derive(Debug, PartialEq, Clone)]
